pyproject.toml
[project]
name = "ontology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.3",
    "openai>=1.64.0",
 "pydantic>=2.10.6",
 "python-dotenv>=1.0.1",
]

[project.scripts]
main = "ontology.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
